crdt v0.1.0 Crdt.LWWSet

A LWW-Set is set that supports removal and insertion of items an arbitrary number of times, but client must provide timestamps for these events. The set can be tuned in a way that it is biased towards adds or deletes, if an add operation occurs at the same timstamp of a remove operation.

Types

bias()

bias() :: :add | :remove

t()

t() :: %Crdt.LWWSet{a_map: value_map(), bias: bias(), r_map: value_map()}

value_map()

value_map() :: %{optional(any()) => any()}

Functions

add(set, item, timestamp)

add(t(), any(), any()) :: t()

Adds item to set at timestamp.

get(set)

get(t()) :: MapSet.t(any())

Returns all items in set.

member?(set, item)

member?(t(), any()) :: boolean()

Returns true if item is a member of set.

merge(s1, s2)

merge(t(), t()) :: t()

Merges s1 and s2.

new(bias)

new(bias()) :: t()

Returns a new, empty LWW-Set.

remove(set, item, timestamp)

remove(t(), any(), any()) :: t()

Adds item to set at timestamp.