crdt v0.1.0 Crdt.MVReg

A MV-Reg is a register that maintains all values when there are conflicting entries. These conflicting entries are detected using Vector Clocks.

Types

t()

t() :: [{any(), Crdt.VectorClock.t()}]

Functions

get(reg)

get(t()) :: MapSet.t(any())

Returns a list of values associated with reg.

merge(r1, r2)

merge(t(), t()) :: t()

Merges r1 and r2.

new()

new() :: t()

Returns a new, empty MV-Reg.

update(reg, id, value)

update(t(), any(), any()) :: t()

Updates reg with value from actor id. It will store all conflicting entries that do have parent Vector Clocks.