crdt v0.1.0 Crdt.VectorClock

A Vector Clock is capable of generating a partial ordering of events in a distributed system and detecting causality vioations. It is essentially a map of actors to their logical clock value.

Types

t()

t() :: %{optional(any()) => non_neg_integer()}

Functions

apply_dot(v, arg)

apply_dot(t(), {any(), non_neg_integer()}) :: t()

Apply a {id, count} to the clock.

concurrent?(v1, v2)

concurrent?(t(), t()) :: boolean()

Returns true if v1 and v2 have diverged.

descends?(v1, v2)

descends?(t(), t()) :: boolean()

Returns true if v1 is a direct descendent of v2. Note that a vector clock is its own descendent.

dominates?(v1, v2)

dominates?(t(), t()) :: boolean()

Returns true if v1 descends v2 and v1 does not equal v2.

forget(v1, v2)

forget(t(), t()) :: t()

Forget any actors in v1 that have smaller or equal counts than the count in v2.

get(v, id)

get(t(), any()) :: non_neg_integer()

Returns the count of id in v if it exists. Else it will return 0.

greatest_lower_bound(v1, v2)

greatest_lower_bound(t(), t()) :: t()

Returns the greatest-lower-bound of v1 and v2.

increment(v, id, value \\ 1)

increment(t(), any(), non_neg_integer()) :: t()

Increments the count of id in v by value.

intersection(v1, v2)

intersection(t(), t()) :: t()

Returns a vector clock of all the common dots in s1 and s2.

merge(v1, v2)

merge(t(), t()) :: t()

Merges v1 and v2.

new()

new() :: t()

Returns a new, empty Vector Clock.